api-tienda-efren/requirements.txt
asgiref==3.3.4
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.2.4
django-heroku==0.3.1
django-rest-swagger==2.2.0
djangorestframework==3.12.4
drf-yasg==1.20.0
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
gunicorn==20.1.0
idna==2.10
inflection==0.5.1
itsdangerous==2.0.1
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
openapi-codec==1.3.2
packaging==20.9
psycopg2==2.9.1
pyparsing==2.4.7
python-dateutil==2.4.2
python-decouple==3.4
python-dotenv==0.17.1
pytz==2021.1
requests==2.25.1
ruamel.yaml==0.17.9
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.16.0
SQLAlchemy==1.4.18
sqlparse==0.4.1
tldextract==1.6
unicodecsv==0.13.0
uritemplate==3.0.1
urllib3==1.26.5
Werkzeug==2.0.1
whitenoise==5.2.0
